requirements.txt
numpy # Don't remove
soundfile # Don't remove
museval
mir_eval
demucs
pyyaml
ml_collections
